feat(reorganizando pastass, corrigindo filtros): pastas reorganizadas, filtros corrigidos
