--- src/components/screens/FeedScreen/index.tsx
@@ -1,6 +1,0 @@
-
-export default function Feed () {
-  return (
-    <div>olá</div>
-  )
-}
