+++ src/components/screens/app/FeedScreen/index.tsx
@@ -1,0 +1,6 @@
+
+export default function FeedScreen () {
+  return (
+    <div>olá</div>
+  )
+}
